[TASK] Change Github hook processing
Until now processing stopped at the first match of known-github-manuals.
Now all lines in known-github-manuals.txt are processed so we can
easily account for branches in Github repositories by creating
a single line for each of the branches.

There now can be empty lines and comment lines (starting with '#')
in the file 'known-github-manuals.txt'.